requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2

# Data processing
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.1.78
scikit-image==0.21.0
scikit-learn==1.3.0
albumentations==1.3.1

# Deep learning
torch==2.0.1
torchvision==0.15.2
timm==0.9.5
pytorch-grad-cam==0.2.1

# Model optimization (optional)
onnx==1.14.1
onnxruntime==1.15.1

# Utilities
tqdm==4.66.1
matplotlib==3.7.3
pillow==10.0.1